requirements.txt
Django==1.7.4
Pillow==2.7.0
argparse==1.2.1
django-crispy-forms==1.4.0
django-filter==0.9.2
django-ipware==0.1.0
django-tables2==0.15.0
django-tables2-reports==0.0.10
psycopg2==2.6
six==1.9.0
wsgiref==0.1.2
